Added actual ending, "completed game"
Game is pretty much complete. May need to come back and fix a couple of issues. Such as the ending animation not being completely fading away. Need to fix the animations of the player when on the wall and floor at the same time. Mouse click not working when game restarts.